+++ src/Linna/Authorization/Authorize.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Linna\Authorization;
 
-use Linna\Authentication\Authenticate;
+use Linna\Authentication\Authentication;
 use Linna\DataMapper\NullDomainObject;
 
 /**
@@ -25,9 +25,9 @@ class Authorize
     protected $permissionMapper;
 
     /**
-     * @var Authenticate Current authentication status
+     * @var Authentication Current authentication status
      */
-    protected $authenticate;
+    protected $authentication;
 
     /**
      * @var int User id
@@ -41,7 +41,7 @@ class Authorize
 
     /**
      * Class Constructor.
-     * <pre><code class="php">use Linna\Auth\Authenticate;
+     * <pre><code class="php">use Linna\Auth\Authentication;
      * use Linna\Auth\Authorize;
      * use Linna\Auth\Password;
      * use Linna\Session\Session;
@@ -52,28 +52,28 @@ class Authorize
      * $password = new Password();
      * $session = new Session();
      *
-     * $authenticate = new Authenticate($session, $password);
+     * $authentication = new Authentication($session, $password);
      * $permissionMapper = new PermissionMapper();
      *
-     * $authorize = new Authorize($authenticate, $permissionMapper);
+     * $authorize = new Authorize($authentication, $permissionMapper);
      * </code></pre>
      *
-     * @param Authenticate              $authenticate
+     * @param Authentication            $authentication
      * @param PermissionMapperInterface $permissionMapper
      */
-    public function __construct(Authenticate $authenticate, PermissionMapperInterface $permissionMapper)
+    public function __construct(Authentication $authentication, PermissionMapperInterface $permissionMapper)
     {
-        $this->authenticate = $authenticate;
+        $this->authentication = $authentication;
         $this->permissionMapper = $permissionMapper;
 
-        $this->userId = $authenticate->getLoginData()['user_id'] ?? 0;
+        $this->userId = $authentication->getLoginData()['user_id'] ?? 0;
 
         $this->hashTable = $permissionMapper->fetchUserPermissionHashTable($this->userId);
     }
 
     /**
      * Check if authenticated user has a permission.
-     * <pre><code class="php">$authorize = new Authorize($authenticate, $permissionMapper);
+     * <pre><code class="php">$authorize = new Authorize($authentication, $permissionMapper);
      *
      * //with this example, the class checks if the authenticated
      * //user has the permission 'update user'.